refactor: Calculator interface 분리
